+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+### Added
+
+- Latency status
+- Latency to the HTML view
+
 ### [2.0.1] 2024-05-31
 
 ### Added